make sure build from maven
