+++ com.swdc.ide.core/build.properties
@@ -9,4 +9,5 @@ bin.includes = META-INF/,\
                plugin.properties,\
                plugin.xml,\
                icons/,\
-               sw.png
+               sw.png,\
+               extlibs/gson-2.7.jar
