+++ com.swdc.ide.core/pom.xml
@@ -1,33 +1,48 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
 
-  <parent>
+	<parent>
 		<groupId>com.swdc.ide</groupId>
 		<artifactId>com.swdc.ide.parent</artifactId>
 		<version>0.1.27-SNAPSHOT</version>
 		<relativePath>../com.swdc.ide.parent</relativePath>
-  </parent>
+	</parent>
 
 
-  <artifactId>com.swdc.ide.core</artifactId>
-  <packaging>eclipse-plugin</packaging>
+	<artifactId>com.swdc.ide.core</artifactId>
+	<packaging>eclipse-plugin</packaging>
 
-  <name>Software.com-core</name>
+	<name>Software.com-core</name>
 
-  <build>
-    <sourceDirectory>src</sourceDirectory>
 
-    <resources>
-      <resource>
-        <directory>${basedir}/src</directory>
-        <includes>
-          <include>**/*.properties</include>
-        </includes>
-      </resource>
-    </resources>
-  </build>
+
+	<build>
+		<sourceDirectory>src</sourceDirectory>
+
+		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.0</version>
+				<configuration>
+					<source>1.8</source>
+					<target>1.8</target>
+				</configuration>
+			</plugin>
+		</plugins>
+		<resources>
+			<resource>
+				<directory>${basedir}/src</directory>
+				<includes>
+					<include>**/*.properties</include>
+				</includes>
+			</resource>
+		</resources>
+	</build>
 
 </project>
