+++ com.swdc.ide.core/src/com/swdc/ide/core/ui/SoftwareCo.java
@@ -438,7 +438,7 @@ public class SoftwareCo {
 				//
 				clientMgr.sendKeystrokeData(keystrokeCount);
 				
-				SoftwareCoUtils.reDisplayStatusMessage();
+				
 			}
 		}
 	}
